+++ src/plugins/saved_objects_management/public/management_section/mount_section.tsx
@@ -104,6 +104,7 @@ export const mountManagementSection = async ({
                   serviceRegistry={serviceRegistry}
                   actionRegistry={pluginStart.actions}
                   columnRegistry={pluginStart.columns}
+                  namespaceRegistry={pluginStart.namespaces}
                   allowedTypes={allowedObjectTypes}
                   setBreadcrumbs={setBreadcrumbs}
                 />