+++ src/plugins/saved_objects_management/server/routes/find.ts
@@ -45,6 +45,9 @@ export const registerFindRoute = (
           perPage: schema.number({ min: 0, defaultValue: 20 }),
           page: schema.number({ min: 0, defaultValue: 1 }),
           type: schema.oneOf([schema.string(), schema.arrayOf(schema.string())]),
+          namespaces: schema.maybe(
+            schema.oneOf([schema.string(), schema.arrayOf(schema.string())])
+          ),
           search: schema.maybe(schema.string()),
           defaultSearchOperator: schema.oneOf([schema.literal('OR'), schema.literal('AND')], {
             defaultValue: 'OR',
@@ -66,6 +69,20 @@ export const registerFindRoute = (
       const managementService = await managementServicePromise;
       const { client } = context.core.savedObjects;
       const searchTypes = Array.isArray(req.query.type) ? req.query.type : [req.query.type];
+      if ('namespaces' in req.query) {
+        const namespacesToInclude = Array.isArray(req.query.namespaces)
+          ? req.query.namespaces
+          : [req.query.namespaces];
+        const searchNamespaces = [];
+        namespacesToInclude.forEach((ns) => {
+          if (ns == null) {
+            searchNamespaces.push('default');
+          } else {
+            searchNamespaces.push(ns);
+          }
+        });
+        req.query.namespaces = searchNamespaces;
+      }
       const includedFields = Array.isArray(req.query.fields)
         ? req.query.fields
         : [req.query.fields];